Changed metagol_sn/4 to return a flattened theory.
Added flatten_snt/3 which does the actual flattening using additional metarules. The rules are called 'snt$<OriginalName>'. The rule is created and asserted on the fly if it is not already in the database. However, only a user:metarule_init/6 is asserted such that this does not change learning behaviour. This is used for pretty printing and program assertion.

Tweaked atomsaslists_to_atoms/2 and add_path_to_body/4 to be able to handle negation necessary for SNTs.